feat: modularize application entry point by extracting routing and root component logic
